Split StyleCorrector from StyleChecker
