+++ lib/policial.rb
@@ -13,7 +13,9 @@ require 'policial/line'
 require 'policial/patch'
 require 'policial/pull_request'
 require 'policial/pull_request_event'
+require 'policial/style_operation'
 require 'policial/style_checker'
+require 'policial/style_corrector'
 require 'policial/linters/base'
 require 'policial/linters/ruby'
 require 'policial/linters/scss'
